context_engine/pyproject.toml
[project]
name = "context-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.55.0",
    "google-generativeai>=0.8.5",
    "jsonpatch>=1.33",
    "langchain>=1.1.3",
    "langchain-core>=1.1.3",
    "langchain-google-genai>=4.0.0",
    "langgraph>=1.0.4",
    "pip>=25.3",
    "psutil>=7.1.3",
    "pynput>=1.8.1",
    "pyperclip>=1.11.0",
    "pywin32>=311",
    "watchdog>=6.0.0",
]
